# ARTEMIS-3033 - implement address tree map for wildcards in place of l…