--- a/artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/Bindings.java
+++ b/artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/Bindings.java
@@ -36,6 +36,8 @@ public interface Bindings extends UnproposalListener {
 
    Binding removeBindingByUniqueName(SimpleString uniqueName);
 
+   SimpleString getName();
+
    void setMessageLoadBalancingType(MessageLoadBalancingType messageLoadBalancingType);
 
    MessageLoadBalancingType getMessageLoadBalancingType();