--- a/artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/impl/AddressImpl.java
+++ b/artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/impl/AddressImpl.java
@@ -16,16 +16,11 @@
  */
 package org.apache.activemq.artemis.core.postoffice.impl;
 
-import java.util.Collection;
-import java.util.Collections;
 import java.util.Set;
 
-import io.netty.util.internal.PlatformDependent;
 import org.apache.activemq.artemis.api.core.SimpleString;
 import org.apache.activemq.artemis.core.config.WildcardConfiguration;
 import org.apache.activemq.artemis.core.postoffice.Address;
-import org.apache.activemq.artemis.utils.collections.ConcurrentHashSet;
-import org.jctools.maps.NonBlockingHashSet;
 
 /**
  * Splits an address string into its hierarchical parts using {@link WildcardConfiguration#getDelimiter()} as delimiter.
@@ -68,31 +63,6 @@ public class AddressImpl implements Address {
    @Override
    public boolean containsWildCard() {
       return containsWildCard;
-   }
-
-   @Override
-   public Collection<Address> getLinkedAddresses() {
-      final Collection<Address> linkedAddresses = this.linkedAddresses;
-      if (linkedAddresses == null) {
-         return Collections.emptySet();
-      }
-      return linkedAddresses;
-   }
-
-   @Override
-   public void addLinkedAddress(final Address address) {
-      if (linkedAddresses == null) {
-         linkedAddresses = PlatformDependent.hasUnsafe() ? new NonBlockingHashSet<>() : new ConcurrentHashSet<>();
-      }
-      linkedAddresses.add(address);
-   }
-
-   @Override
-   public void removeLinkedAddress(final Address actualAddress) {
-      if (linkedAddresses == null) {
-         return;
-      }
-      linkedAddresses.remove(actualAddress);
    }
 
    /**