--- a/artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/impl/BindingsImpl.java
+++ b/artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/impl/BindingsImpl.java
@@ -86,6 +86,7 @@ public final class BindingsImpl implements Bindings {
       this.name = name;
    }
 
+   @Override
    public SimpleString getName() {
       return name;
    }