@franz1981 2020-12-16
I suggest to run the benchs pre/post this PR, possibly by setting a fixed CPU Freq on an idle machine