@clebertsuconic 2020-12-16
There's a functionality of versioning on Bindings... Can you make sure that functionality is not affected when a Binding is updated?

Say, you create a producer.. send a message...


Then you update the bindings...


Send another message with the previous producer...


The output should still be valid... on the updated address..


Can you add a test doing that?