@gtully 2020-12-16
@clebertsuconic the second send should go to any new binding. I can verify that, but the changes go through bindings.addBinding so that will invalidate/increment the version.